requirements.txt
# Core scientific computing stack
numpy==1.24.4
pandas==2.0.3
scikit_learn==1.3.2

# PyTorch & GNN frameworks (no version constraints)
torch
dgl
scipy
torch_scatter
torch_sparse
torch_geometric

# Other related libraries
ogb==1.3.6
tqdm==4.66.4
PyYAML==6.0.1
GPUtil==1.4.0
networkx==3.1

GraphRicciCurvature==0.5.3.2
